refactor(graduate): thesis 관리자 페이지 page 레이어로 이동 및 컴포넌트 이름 변경
