--- apps/graduate/src/features/thesis-management/components/index.ts
@@ -1,1 +1,0 @@
-export { default as ThesisManagement } from './ThesisManagement';
