+++ apps/graduate/src/pages/admin/thesis/index.ts
@@ -1,0 +1,1 @@
+export { default as ThesisAdminPage } from './ui/ThesisAdminPage';
