+++ apps/graduate/src/pages/admin/thesis/mock/mockRows.ts
@@ -1,4 +1,4 @@
-import type { ThesisRow } from '~/features/thesis-management/types/rows';
+import type { ThesisRow } from '../types/rows';
 
 const names = [
   '서진규',
