+++ apps/graduate/src/routes/_afterLogin/thesis.tsx
@@ -1,14 +1,13 @@
 import { createFileRoute } from '@tanstack/react-router';
 
-import { ThesisManagement } from '~/features/thesis-management/components';
-
+import { ThesisAdminPage } from '~/pages/admin/thesis';
 import { ThesisPage } from '~/pages/client/thesis';
 
 export const Route = createFileRoute('/_afterLogin/thesis')({
   component: () => {
     const { auth } = Route.useRouteContext();
     if (auth.isAdmin) {
-      return <ThesisManagement />;
+      return <ThesisAdminPage />;
     }
     return <ThesisPage />;
   },
